New display folder, move existing objects
